+++ src/Event.h
@@ -7,10 +7,10 @@ using namespace std;
 class Event : public AbstractEvent {
  int8_t pitch;
 public:
- Event(Time,Time,int);
+ Event(Time,Time,int8_t);
  Event(Time,Time);
  Event* copy();
  int getPitch();
- void setPitch(int);
+ void setPitch(int8_t);
 };
 #endif // EVENT_H
